ML/requirements.txt
torch
torchvision
facenet-pytorch
fastapi
uvicorn
numpy
pillow
opencv-python
